Added initialization of CustomAsteroidData
Data will now be moved from scenario when an asteroid is first loaded. Due to unknown limitations, the action needs to be triggered on part load, not on vessel load, which could cause trouble if there are multiple asteroid parts on the same vessel.

Also changed SpawnCatcher trigger from `onVesselCreate` to `onNewVesselCreated`, which should minimize the risk of false positives.